+++ core/src/main/kotlin/terapps/factoryplanner/core/services/components/Destination.kt
@@ -2,6 +2,6 @@ package terapps.factoryplanner.core.services.components
 
 data class Destination(
         val factorySiteId: String,
-        val targetAmountPerCycle: Float,
+        val targetAmountPerCycle: Double,
 ) {
 }
